+++ librlox/src/parser/mod.rs
@@ -1,4 +1,5 @@
 pub mod expression;
+pub mod expression_parser;
 pub mod statement;
 pub mod statement_parser;
